Code/requirements.txt
google-cloud-storage
google-cloud-speech
pydub
moviepy
fastapi
uvicorn
python-multipart
pydantic==1.8.2
moviepy==1.0.3
